# Write raw tags directly to buffer