--- a/ext/liquid_c/raw.c
+++ b/ext/liquid_c/raw.c
@@ -82,6 +82,8 @@ static VALUE raw_parse_method(VALUE self, VALUE tokens)
             body_len += match.body_len;
             VALUE body_str = rb_enc_str_new(body, body_len, utf8_encoding);
             rb_ivar_set(self, id_ivar_body, body_str);
+            tokenizer->raw_tag_body = RSTRING_PTR(body_str);
+            tokenizer->raw_tag_body_len = (unsigned int)body_len;
             return Qnil;
         }