Write raw tags directly to the instructions buffer as an `OP_WRITE_RAW` instruction. This is an alternative to #134.

# Benchmarks

There is a small performance hit during parsing, but a slight performance improvement during rendering in the microbenchmark.

Master:

```
               parse    105.637k (± 6.9%) i/s -    535.392k in   5.092843s
              render    206.248k (± 8.5%) i/s -      1.036M in   5.058648s
      parse & render     57.954k (± 6.6%) i/s -    292.464k in   5.069258s
```

This branch:

```
               parse     97.280k (± 3.3%) i/s -    486.423k in   5.006012s
              render    225.835k (± 6.2%) i/s -      1.126M in   5.009018s
      parse & render     53.268k (± 8.9%) i/s -    268.041k in   5.077031s
```

Script:

```ruby
require "benchmark/ips"
require "liquid"
require "liquid/c"

source = <<~LIQUID
  {% raw %}
    {% test %}
    hello world
    {% endtest %}
  {% endraw %}
LIQUID

template = Liquid::Template.parse(source)

Benchmark.ips do |x|
  x.report("parse") do |times|
    times.times do
      Liquid::Template.parse(source)
    end
  end

  x.report("render") do |times|
    times.times do
      template.render!
    end
  end

  x.report("parse & render") do |times|
    times.times do
      Liquid::Template.parse(source).render!
    end
  end
end
```